+++ src/java/SergeantCohort/ElectionContext.java
@@ -9,15 +9,15 @@ import java.util.HashSet;
 public class ElectionContext
 {
     public final long last_view_number;
-    public final int voting_for_cohort_id;
+    public final long voting_for_cohort_id;
 
     /**
        Only for those trying to become leaders themselves.
      */
-    public final Set<Integer> votes_received_set = new HashSet<Integer>();
+    public final Set<Long> votes_received_set = new HashSet<Long>();
     
     
-    public ElectionContext(long last_view_number,int voting_for_cohort_id)
+    public ElectionContext(long last_view_number,long voting_for_cohort_id)
     {
         this.last_view_number = last_view_number;
         this.voting_for_cohort_id = voting_for_cohort_id;
